anime_object_detection/requirements.txt
gradio>=4.44.0
numpy
pillow
onnxruntime
huggingface_hub
scikit-image
pandas
opencv-python>=4.6.0
hbutils>=0.9.0
dghs-imgutils>=0.5.0
git+https://github.com/deepghs/imgutils.git
httpx
